15000 persons, 150 churches ! It works !

+++ src/GermBundle/Command/RemoveAllDataCommand.php
@@ -7,12 +7,12 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 
-class EmptyAllDataCommand extends ContainerAwareCommand
+class RemoveAllDataCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
         $this
-            ->setName('germ:empty_all_data')
+            ->setName('germ:remove-all-data')
             ->setDescription('Remove all existing data')
             ->setHelp('Remove churches and persons. Not allowed on production. For testing and work with data for developing without real data.')
         ;
@@ -20,9 +20,18 @@ class EmptyAllDataCommand extends ContainerAwareCommand
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        if ($this->getContainer()->getParameter("kernel.environment") == 'prod') {
+            throw new \Exception("This command cannot be executed on PROD environment", 1);
+        }
+
         /** @var Session $pommSession */
         $pommSession = $this->getContainer()->get('pomm')['germ'];
         $pommSession->getConnection()->executeAnonymousQuery(<<<SQL
+            DELETE FROM event.assignation;
+            DELETE FROM event.docket;
+            DELETE FROM event.event;
+            DELETE FROM event.event_type;
+            DELETE FROM event.location;
             DELETE FROM person.account;
             DELETE FROM person.person_church;
             DELETE FROM person.person;
